Use JSON instead of ByteArray for network protocol
The server now accepts strings, which are in a valid JSON format. All
other strings or messages are dropped.
This enables other clients than Java clients to communicate with the
ConAir network and the server can now handle any packet (as long it is
JSON based) and broadcast them to clients.